Cast model types to int32

+++ src/sequifier/helpers.py
@@ -8,7 +8,7 @@ def numpy_to_pytorch(data, seq_length, device):
 
     sequence = (
         tensor(data[[str(c) for c in range(seq_length, 0, -1)]].values)
-        .to(torch.int64)
+        .to(torch.int32)
         .to(device)
     )
